Fix bug in Python codemods
The way the codemod was written had it so that if once it fired
on a node, it would fire on all subsequent nodes. This is clearly
not ideal! This fixes that, adds some tests for it, and adds
some performance improvements for issues that were being masked
by this.

Fixes #12